fix : 문장 저장 시 WhiteSpace 지우고 저장

+++ src/main/java/teamverpic/verpicbackend/domain/analysis/domain/Script.java
@@ -18,7 +18,7 @@ public class Script {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long scriptId;
-    private int wpm;
+    private double wpm;
 
     @OneToOne(mappedBy = "script")
     private AudioFile audioFile;
